move component to molecules and rename
